refactor(Cars): 패키징 수정
