+++ src/test/java/step3and4/car/CarsTest.java
@@ -1,9 +1,7 @@
-package step3and4.cars;
+package step3and4.car;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import step3and4.car.Car;
-import step3and4.car.Position;
 import step3and4.number.Number;
 import step3and4.number.RandomNumber;
 import step3and4.output.OutputView;
